test: reorganize modules and cover invalid inputs
