+++ sitegen/src/parser.rs
@@ -113,3 +113,60 @@ pub fn read_roles() -> BTreeMap<String, String> {
             ])
         })
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn parses_english_months() {
+        let months = [
+            ("January", 1),
+            ("February", 2),
+            ("March", 3),
+            ("April", 4),
+            ("May", 5),
+            ("June", 6),
+            ("July", 7),
+            ("August", 8),
+            ("September", 9),
+            ("October", 10),
+            ("November", 11),
+            ("December", 12),
+        ];
+        for (name, number) in months {
+            assert_eq!(month_from_en(name), Some(number));
+        }
+    }
+
+    #[test]
+    fn parses_russian_months() {
+        let months = [
+            ("Январь", 1),
+            ("Февраль", 2),
+            ("Март", 3),
+            ("Апрель", 4),
+            ("Май", 5),
+            ("Июнь", 6),
+            ("Июль", 7),
+            ("Август", 8),
+            ("Сентябрь", 9),
+            ("Октябрь", 10),
+            ("Ноябрь", 11),
+            ("Декабрь", 12),
+        ];
+        for (name, number) in months {
+            assert_eq!(month_from_ru(name), Some(number));
+        }
+    }
+
+    #[test]
+    fn invalid_english_month_returns_none() {
+        assert_eq!(month_from_en("Foo"), None);
+    }
+
+    #[test]
+    fn invalid_russian_month_returns_none() {
+        assert_eq!(month_from_ru("Бар"), None);
+    }
+}
