+++ sitegen/tests/inline_start.rs
@@ -1,0 +1,35 @@
+use sitegen::{read_inline_start, InlineStartError};
+use std::env;
+use std::fs;
+
+#[test]
+fn reads_inline_start_from_markdown() {
+    let dir = tempfile::tempdir().expect("temp dir");
+    let original = env::current_dir().unwrap();
+    env::set_current_dir(dir.path()).unwrap();
+    fs::write("cv.md", "* March 2024 – Present").unwrap();
+    let result = read_inline_start();
+    env::set_current_dir(original).unwrap();
+    assert_eq!(result.unwrap(), (2024, 3));
+}
+
+#[test]
+fn read_inline_start_returns_error_for_invalid_month() {
+    let dir = tempfile::tempdir().expect("temp dir");
+    let original = env::current_dir().unwrap();
+    env::set_current_dir(dir.path()).unwrap();
+    fs::write("cv.md", "* Foo 2024 – Present").unwrap();
+    let result = read_inline_start();
+    env::set_current_dir(original).unwrap();
+    assert!(matches!(result, Err(InlineStartError::Parse)));
+}
+
+#[test]
+fn read_inline_start_returns_error_when_file_missing() {
+    let dir = tempfile::tempdir().expect("temp dir");
+    let original = env::current_dir().unwrap();
+    env::set_current_dir(dir.path()).unwrap();
+    let result = read_inline_start();
+    env::set_current_dir(original).unwrap();
+    assert!(matches!(result, Err(InlineStartError::Io(_))));
+}
